+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add support for JMX options configuration of all Kafka Connect (KC, KC2SI, MM2)
 * Updated Cruise Control to version 2.5.32
 * Fix Cruise Control crash loop when updating container configurations
+* Configure extenal logging `ConfigMap` name and key.
 
 ### Deprecations and removals
 
@@ -16,6 +17,22 @@
   * `operator.strimzi.io/generation` (used internally only - replaced by `strimzi.io/generation`)
   * `operator.strimzi.io/delete-pod-and-pvc` (use `strimzi.io/delete-pod-and-pvc` instead)
   * `operator.strimzi.io/manual-rolling-update` (use `strimzi.io/manual-rolling-update` instead)
+* External logging configuration has changed. `spec.logging.name` is deprecated. Moved to `spec.logging.valueFrom.configMapKeyRef.name`. Key in the `ConfigMap` is configurable via `spec.logging.valueFrom.configMapKeyRef.key`.
+  * from
+  ```
+  logging:
+    type: external
+      name: my-config-map
+  ```
+  * to
+  ```
+  logging:
+    type: external
+    valueFrom:
+      configMapKeyRef:
+        name: my-config-map
+        key: my-key
+  ``` 
 
 ## 0.21.0